requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
django-cors-headers>=4.3,<5.0
gunicorn>=22.0,<23.0
drf-spectacular>=0.27
whitenoise>=6.0
djangorestframework-simplejwt==5.3.1
django-import-export==4.1.1
